Work around for asdf code change
I am about to deliver a change to asdf which will improve the way that
date tagging is handled. The result of the change is that complex
formats, such as dates, are stored as tagged strings. This is not a
problem, and in fact is an improvement, as that is the format that
both fits and serialized asdf want to see. However, the current
version of asdf stores dates as dates and model_base.py contains a
patch to work around the problem it causes. Because asdf and datamodel
are separatepackages I needed to add some code so that it will work
with both the old and about to be delivered version of asdf, basically
just an if statement.