Stop incorrect promotion of FITS_Rec object to ndarray
Part of the process of reading any fits hdu is comparing the data to a
schema and copying the data to the correct data type when the type of
the data and schema do not match. Function util.gentle_asarray does
the check and possible copy. However, it was calling numpy.asarray, which
always produces an ndarray instead of numpy.asanyarray, which produces
a new object of the same type as the object being copied. This causes
problems when reading fits tables, as the type FITS_Rec subclasses
ndarray and not copying the data as a FITS_Rec loses access to its
methods and fields. Changing the calls to numpy.asanyarray resolves
the problem.