Update for handling NIRSpec data.
The NIRSpec team has proposed changes to the flat_field step to handle
the dependence of the flat field on wavelength.  This version reads and
uses reference files that are roughly the same as dummy files delivered
by the NIRSpec team in June.

This is a work in progress, and there are still many print statements
for testing and debugging.