Check value used to intialize datamodels and check shape
Two changes were made to datamodels in response to reported errors
when using it.

The first change was to model_base.py. The __init__ method takes a
parameter used to intialize the model, which may be a filename or
various other things. If the parameter matches none of the possible
choices, it used to fall off the end of the if statement and not
initialize the model properly, leading to a fatal error. The new code
adds an else branch to the if which raises a ValueError saying that
the parameter is not of the right type.

The second change is in the code in properties,py which creates a
default array when none is supplied. To create the array it first
generates an empty array of the proper shape and then fills it with
the default value. It gets the shape from the schema, but when it was
not defined there, the code threw an error. The fix sets the shape to
zero when it is not defined in the schema.