Add resolver as an optional parameter when creating a new datamodel
The resolver maps uri's contained in references in a schema to
filenames. The default list of mappings is hard coded. However one can
create different mappings by creating a Resolver object. Previously
there was no way to use an alternate resolver when creating a
datamodel. This change adds resolver as an optional parameter to
__init__. This allows the programmer to provide an alternate resolver
and reference schema files in directories other than the defaults.