Retrieve bintable column names from table instead of datamodels schema
While the bintable code makes some attempt to make column access
independent of the column name case, the code is layered on numpy
records, where access is case dependent. In practice, you must supply
column names in the correct case. The function gentle_asarray coerces
the type of data read from a model into the type given in the
schema. However, in the case of numpy records, the field name is a
part of the datatype. This leads to erroneious results if the case of
the column name mentioned in the schema disagress with the case of the
name in the binary table. The fix replaces the column names read from
the schema with the column names from binary table.

It looks from the test case for gentle_asarray that the intent was to
coerce the column names in the table to match the names in the schema.
However, there is no easy way to do this and no interace to accomplish
it, so this behavior is not supported, if it was even ever intended.