Fix problems uncovered by badly formed test image
Three potential problems were found and fixed. None should happen with
corectly formatted images.

The first change was to datamodels/__init__.py. There is a test on the
dimesnions of the image in open in order to determine the right class
to use. The change uses an explicit extension version in case there is
more than one SCI extension in an image.

The second change is to _cast in datamodels/properties.py. If the
object to be cast is a NDArrayType, then it is lazy and must first be
instantiated with a call to _make_array before calling gentle_asarray.

The third change is to gentle_asarray in datamodels/util.py. The new
code traps any error when calling numpy.asarray on an object that is
not a numpy.ndarray and reports the object as having a bad type.