Adjust fits record names to match schema
Function gentle_asarray coerces datatypes to match the type specified
in the schema. This can lead to a mismatch with the column names in
the table. Something has to give and previously I overrode the schema
column names with the fits record column names. This caused
regressions in the overnight tests. This fix flips the coercion and
updates the fits record column names to match the schema column
names. I hadn't done this before because I though the change was more
difficult.