Update numpy dtype record to match names read from fits table
The function gentle_asarray coerces records read from fits files to
the type mentioned in the corresponding schema file. This can lead to
problems with fits tables, as in the cose of numpy records, the
coercion to the schema types includess coercion to the schema field
names. The new names may not match the case of the names stored in the
fits table and where there is a mismatch, case insensitive table
column matching fails. The change to gentle_asarray in
datamodels/util.py makes sure the field names in the numpy record
matches the names in the fits table, including the case of the name.

Because the case coercion is no longer done, the test of gentle_asarry
in datamodels/tests/test_storage.py had to be updated to remove the
test for case coercion.