Remove resolver as a parameter, use extension instead to add new url mappings
When Nadia saw my update which added resolver ss a parameter to
datamodels/model_base.py, she said it was redundant and I could pass
an extension to update the url_mapping instead. And that this was the
intended method to do this. I reviewed the code and saw she was right,
so I have removed resolver as a parameter.

But this presented another problem, because when load_schema is called
from __init__, it needs to receive a resolver in order for it to be
able to use the updated mappings. To generate this mapping, I had to
put the code that handles extensions before the call to load_schema
and I had to generate the resolver for load_schema from those
extensions. And after a little research, I have maded the needed lines
to do that.