Add url mapping for datamodels schemas
Up to now there has been no way for other packages to refer to the
schemas in the datamodels package. The miri team requested that the
url http://jwst.stsc.edu/schemas be added, so their schemas could have
references to the datamodels schemas. This commit adds the new class
BaseExtension in datamodels/extension.py and modifies the list of
default extensions, jwst_extensions, in datamodels/model_base.py to
include it.