Add instance method to datamodel properties node
The Node object is a bundle made from a value, its schema, and its
context. Sometimes you want to retrieve the value from the node, if,
for example, you would like to copy it. The new method instance has
been added to the Node class and it returns the value. It is invoked
as:

    val = node.instance

where node is the object you wish to fetch the value from. I have also
rewritten the internal function _unmake_node to use this new method.