Correct argument list for copy
The DataModel class implements a copy method and the __copy__ and
__deepcopy__ methods are alaised to it. These two methods are called
by the copy module and the copy.deepcopy method requries that
__deepcopy__ takes a memo argument. When copy.deepcopy is called on a
DataModel object, it fails because of the missing argument. The fix
adds a keyword argument, memo=None, to the copy method and passes that
argument to the invocation of copy.deepcopy that it contains.