Change asarray to asanyarray in _fits_array_writer
Binary tables are stored in memory as a type derived from numpy
arrays. One has to be careful when copying them to use asanyarray
instead of asarray so the`copy has the same type and does not lose the
methods and fields of the base type. In this case, the _coldefs
attribute, which contains the attributes of a column, was lost, making
it impossible to update a column's attributes with the datamodels
code.