Change asarray to asanyarray in _fits_array_writer
Binary tables are stored in memory as FitsRec objects, a type derived
from numpy arrays. One has to be careful when copying them to use
asanyarray instead of asarray so the copy has the same type as the
original and does not lose the methods and fields of the base type.
In this case, the _coldefs attribute, which contains the attributes of
the columns in a table, was being lost, making it impossible to update
the table column attributes with the datamodels code.

This is the same fix submitted yesterday minus several changes that
were mistakenly placed on the branch.