Move test for pass_invalid_values into Node._validate
The flag pass_invalid_values was moved from __setattr__ into _validate
so that all methods which update the asdf tree are treated the
same. The flag overrides validation errors that are triggered when the
asdf tree is updated. However, if the results are saved to a file a
validation error is still thrown. There is no simple way to override
validation errors that happen during a save.

The change was placed in this branch because validation in append was
previously commented out. Presumably whoever did that had a good
reason and this change will permit the user to turn off the validation
when they wish.