Include upper extraction limit; Python 3 round function.
When specifying extraction limits using `xstart` and `xstop` (or
`ystart` and `ystop`), the upper limit was not actually included in
the extraction region, but the documentation referred to `xstop` as
"last pixel in X", which implies that `xstop` should have been
included.  The code was changed to be consistent with the
documentation.  But also change the documentation to make it more
clear.

The use of the `round` function led to differences in the size of
the extraction region, when using Python 3 vs Python 2.