Update TUNIT when column units are changed
It was reported that the TUNIT keyword was not being updated when a
table was opened using the datamodels package and a column's units was
changed. The problem was in fits_support._fits_array_loader. That code
will convert the datatypes of the table columns when they do not agree
with the schema. That conversion involves a copy and the copy loses
the listeners which update the header keywords when a change is made
to a column's attributes. Code was added to copy the listeners from
the original data to the copy of the data so that column units are
updated.