Add ReferencefileModel class
The new class uses referencefile.schema.yaml as its schema, but
otherwise is identical to DataModel, the base class. The generic open
in util.py has been modified to open a file as a ReferencefileModel if
it has the REFTYPE keyword in the primary header.

The purpuse of these changes is to make reference file specific
metadata, such as REFTYPE and the USEAFTER date, available in the
metadata when a file has been opened with the generic open.