Move core schema into referencefile schema
Each individual reference file schema used to include both the core
schema and the referencefile schema. I moved include for the core
schema into the referencefile schema and deleted it from the
individual reference file schemas.

This change was needed to make the ReferencefileModel class work, as
the Datamodels code assumes that all classes include the metadsts in
the core schema.