Nod offset; xstart, xstop, etc; independent variable
The nod offset correction was not always being applied correctly (though
it was OK for the case that `src_coeff` or `bkg_coeff` was specified).
See #549.

If the source extraction limits are specified by `xstart`, `xstop`,
`ystart`, `ystop`, those values are supposed to be integers and inclusive
limits.  In some places, the code was treating the upper limits as
exclusive.  See #550.

The independent variable for polynomial functions was "wavelength"; this
has been changed to "pixel".  See #551.