change API of make_output_wcs() from wcs list to model list
The original implementation of this function accepted a list of
input models (or a model container) because it assumed that the
domain of the wcs objects was not populated.  And it turns out
this is true for the current wcs objects created be assign_wcs
for NIRCam.  No domain.  Meanwhile our mocked-up test data has
domains assigned, which is why our tests didn't fail.  Changing
the API back so that current NIRCam data works, until we fix
assign_wcs to add domains.