Rewrite of twopoint_difference
This is a significant rewrite to fix bugs and improve performance. The
big speed up is due to only sorting the values once and using the
sorted index for both the median calculation and finding the outlier.
Also, this method allows the exclusion of the largest difference from
the median calculation as was specified in the Anderson et al. paper.
There is also a new test routine that can be used to verify the changes
and demonstrate the bugs in the previous version.