Add image based reference file types
This commit adds the ReferenceImageModel, ReferenceCubeModel, and
ReferenceQuadModel classes, which represent generic reference files
whose dimensions are 2, 3, or 4 respectively. The datamodels.open code
has been modified to open reference files using these classes if they
have a SCI extension.

The purpose is to avoid regressions in existing code, which previously
opened these files using the ImageModel and so on classes, as the new
classes are compatible with them.