Move ramp file check in datamodels.open into new function
For the sake of backwards compatibility the test to see if a fits file
is a ramp file need to be done ahead of the test for a generic
reference file. So I have moved the tests into _class_from_ramp_type
and added a call to he new function to datamodels.open.