Remove warning messages from datamodels test output
Datamodels was throwing a lot of warning messages when running
tests because of invalid data. This was for two reasons. First,
the test files contained invalid header keyword values (the
majority of cases.) Second, because some tests set invalid
values (the minority.) This commit fixes the header keyword
values to validate against the schema and and modifies the tests
to set valid values where possible. (Some tests are to check the
code when setting invalid values.)