Determine file type before doing generic open
Previously the generic open (open in datamodels/util.py) first tried
to open a file as a container and if that failed, opened it as a fits
file. That created an undesirable dependency on the internals of
association lists, which is what containers are under the hood. The
new code opens the file and reads the first 5 characters to look for
the magic string that starts asdf and fits files. If no magic string
is found, the file is assumed to be a container. After determining the
file type, each type of file (asdf, container, or fits) is opened in
the appropriate way. Previously there was no code to handle asdf files
in the generic open, so this has been added. The code to determine the
file type has been placed in a new file, filetype.py.