Keyword was being compared with lower-case value; see issue #710.
For IFU data, keyword SRCTYPE is read to determine whether the target
is a point source or extended.  The keyword values were assumed to be
either "point" or "extended"; they are, but upper case.

For an extended source with the IFU, background subtraction is not
supposed to be done.  The code now explicitly turns off background
subtraction for IFU data.