WIP Continuing work on persistence; see issue #595.
The primary header of the output file for the number of filled traps
is now updated based on the input file.  This includes, in particular,
the start and end times of the exposure.  The FILENAME keyword will be
further modified to include a dummy suffix, so that the actual name of
the output file will include the full input file name but with the
dummy suffix replaced by "trapsfilled".