Add support for fitsblender.blendheaders
New code is needed so that datamodels can support fitsblender. The
issue is that fitsblender generates a new hdu containing a binary
table, but the columns of this table vary. The new code generates the
table and the schema that datamodels needs to validate it. It adds a
new dunction to fits_support, from_fits_hdu, which updates a numpy
array which can be added to an existing datamodel.