Add support for astropy.io.registry to datamodels
Astropy.io.registry is a way to provide generic io to different data
formats used with astropy. For each data format a three functions are
registered to read, write, and identify a file as having that
format. The code to register the three functions was added to
__init__.py. The read and write functions were added to util.py and
wrap the open function and save method respectively. The identify
function was added to filetype.py, it calls the check function. A few
lines of glue code were added to model_base.py.