Revise validation error handling in datamodels
Several previous changes to error handling left the code in an
inconsistent sate because they were done in a piecework fashion. All
the validayion error handling has been examined and reworked. There is
a new keyword parameter to DataModel in model_base.py,
warn_if_invalid. If it is set to True (the default) a warning message
is generated. If it is set to False, the message is sent to the log
instead, which by default is the bit bucket. The warning messages all
are of type ValidaionWarning, a new class of Warning added to
properties.py. All warning messages contain the string "is not valid"
so code catching the warning can match on that string as well.