Change error handling in datamodels
Error handling is now controlled by two keyword parameters in the
__init__ method, error_if_invalid and pass_invalid_values. Both are
false by default.

If error_if_invalid is true, datamodels throws a ValidationError if a
fits file does not validate against the schema. It raises a
ValidationWarning if error_if_invalid is false. To convert the warning
message to a log file message, call
"logging.captureWarnings(True)". Datamodels will delay throwing an
error when reading a fits file until all the data is checked so that a
mesaage is printed for all problems with validation. Howerever, an
error will be generated whenever a value is modified after reading, as
the code has no way of knowing how many changes will be made.

If pass_invalid_values is true, datamodels will permit changes to the
files which do not validate. If it is false, invalid changes will be
rolled back.  An error will be generated even if pass_invalid_values
is true if error_if_invalid is also true, the keywords are independent
of one another.

The default case, where both keyword parameters are false, is most
likely what you want. Invalid changes generate a warning mmessage and
the change is rolled back.