Revise and rationalize error handling in datamodels
In its latest iteration, errors are controlled by one keyword,
pass_invalid_value, or its associated environment variable,
PASS_INVALID_VALUE. If it is false, datamodels will throw an
error if a model does not validate against the svhema, but now
the error will not be thrown until the entire file is read,
allowing the user to see a complete list of all the invalid data
via warnings. If pass_invalid_values is true, the same warnings
are generated, but no errors are thrown. If a value does not
validate, it is not placed in the model when reading and any
changes made afterwards are rolled back. This is because the
whole model is validated against the schema and allowing invalid
values into it produces false positives when further values are
checked.