Fix overrides of distorion file units
The DistortionModel type in datamodels is implemented in
distortion.py.  It allows the user to override the units of the
distortion model through the optional parameters input_units and
output_units. The tests which determine whether these optional
parameters are set was incorrect and the result was both units were
set to None. Fixing the tests resolved the problem.