Set wcs to None when naxis is zero in header
ipython shows a warning message when tab is used to auto-complete the list
of attributes of a datamodel. The warning message comes from the wcs code and
says:

The WCS transformation has more axes (2) than the image it is
associated with (0)

The problem is that the code is trying to create a wcs when naxis in
the primary heade is zero. I have modified get_fits_wcs so that it
sets the wcs to None when naxis is zero. This eliminates the warning
message.