Remove _tag_values from datamodels/fits_support.py
When setting an attribute of a datamodel or writing the datamodel out
as a fits file, the code validates the datamodel against its
schema. This schema can contain tag fields. When validating a tag
field, the code checks for equality between the value of the tag field
and the value of the _tag attribute of the corresponding item in the
datamodel. You may ask, what is this _tag attribute and how is it set?
Previously it was set by calling tag_object in asdf/tagged.py from
_cast in datamodels/properties.py and _tag_values in
datamodels/fits_support.py.

But, conscientiously tagging our data values was a mistake, as the
validation code, in validate_tag in asdf/schema.py, skips the check if
the _tag attribute is undefined. Since this check has no value for our
use cases, we are creating needless problems by trying to set
it. @nden has already removed the code in
datamodels/propertie.py. This removes the code in
datamodels/fits_support.py.