Update test that fails, but shouldn't
A test in jwst.datamodels was reporting a failure. The values reported
were ('string', 1., 2. 3., 4.) and ('string', 1.0, 2.0, 3.0, 4.0). The
test should be a success, but was a failure because of how the two
records were compared. I wrote a  small function that compares the two
records element by element and now the test passes.