Make ModelContainer.save consistent with DataModel.save
The `save` method gets passed a filename.  This gets ignored for
ModelContainer, and `meta.filename` is used to write out the file.
This change is needed as Step.save_model method needs the same
API for all DataModels that are returned from a step.

Also, add a filter to suppress "Card is too long" warnings from
`astropoy.io.fits` when writing out a DataModel.  These warning
messages are caused by our schemas containing descriptions that
are too long for the FITS standard in some cases, and in others
by keyword value strings that are very long.