Modify the arguments passed to Step post-hooks
Previously, the result of the parent Step was passed directly to the
post hooks. This commit modifies to the calls to the post hooks to
always pass a list. If the parent Step already passes a list, no
modification is made. However, if the parent Step returns a single
item, that item is placed into a list, and this is passed to the post
hooks.

This allows post-hooks to modify the results.