Close hdulist if it was opened by datamodels.util.open
Several users have reported problems with warning messages from
astropy.io.fits about trying to access an hdu after it was closed. I
looked at the code and found two places where the handling of hdus can
be improved.

In util.open a fits file is opened as a hdulist so the code can
determine which class to use when opening a file. If the fits file is
opened as a hdulist, as opposed to the hdulist being passed as an
argument, it is now closed before exiting util.open.

In model_base.__init__ fits files are opened as a hdulist. The data in
the hdulist is then converted to an asdf file. The hdulist is placed
in a list of objects to close when the datamodel is closed. The code
that adds the hdulist to that list was supposed to be inside a try
block, in case the open failed. Because of incorrect indentation, it
was not. So I changed the indentation to move it into the try block.

I'm not certain that these changes will fix the reported problems, but
since the code was in error, it seemed to be a good attempt at a fix..