Rewrite code that chooses to open file as asdf or fits
The previous code first tried to open the file as asdf and if that
failed, tried to open as fits. The new code looks for the text string
at the start of the file and uses it to identify the file as asdf or
fits. The previous method tied the code too tightly to internal
details of the implementation of asdf.