Change code to close hdulist in datamodels.open
Datamodels.open opens a fits file as a hdulist in order to examine the
header andshape and determine which class to use to open the
file. Since it opens the hdulist, it should also close it. Previously
I had the open function close the hdulist, which was wrong. (What was
I thinking?) The code now appends the hdulist to the list of files to
close when closing the datamodel.