Add iterators to object nodes in datamodels
An iterator was defined on the datamodels meta node as part of the
effort to provide compatibiliy with the astropy NDData
interface. Because having a way to iterate over nodes in a datamodel
is generally useful, the code for the iterator was moved into the
parent of MetaNode, ObjectNode. The iterator walks the tree and
returns keys from subtrees as dotted names,
e.g. "meta.instrument.name".

In addition a new method was added to ObjectNode, items(). It is a
generator which returns a (key, value) tuple with each invocation. It
invokes the new iterator and so the key in the tuple has the same
format.