Change the model container iterator to open one model at a time
The iterator that is part of the model container has been rewritten so
that at each iteration the current model is opened and the previous
model is closed. To support this, the code in model container has been
changed so that models are not opened when the container is created
and random access checks if the model is open and opens it if it is
not. The purpose of this change is to reduce the memory footprint of
code that uses model containers.