Revise ModelContainer iterator to work with existing code
The new ModelContainer iterator opens each model file if it is not yet
open and closes it on the next iterarion. This revision does not close
models that the iterator hasn't opened itself, which allows it to work
with existing code. The code to initialize a container from an
association has been revised so that it does not open the models. This
allows it to work with the iterator and provides the best improvement
over previous code without any revision to it.