Extract filename from hdulist for storage in metadata
The current version of DataModel.__init__ only extracts the filename
from the imput if the input is a string. This version also extracts
the filename if the input is an hdulist. In that case it calls
fileinfo to get the filename from the hdulist. This change was needed
because in the most usual case the generic open function passes an
hdulist to DataModel.__init__, so previously the filename was not
being updated in the metadata.