Remove asn from list of filetypes checked when file does not exist
The new code in the filetype check looks at the extension if the file
does not exist. I've removed "asn" from the list of filetypes because
association files do not have a specific extension assigned to them.
This is not going to be a problem because while the datamodels code
reads associations, it does not write them and will not for the
foreseable future.