A few more changes.
persistence.cfg was modified to change the name from "Persistence" to
"persistence", and the command-line arguments were added.

persistence.py was modified to account for decays during resets,
to handle the case of only one group per integration, and to fix a bug
in the subtraction of decays from the traps-filled map for the case that
the data are a subarray.

See JWST issue #595.