Make models_grouped eager
The method models_grouped in ModelContainer returns a list of list of
models. So it should not use the ModelContainer iterator, instaead it
now opens each model individually, which results in the models staying
open.

In addition, the test cases for models in test_models.py call the
fixture container. It adds metadata to the test files, looping over
each. Because the metadata needs to persist over the life of the
tests, the models need to stay open. Accessing each model through
__getitem__ accomplishes this, as random access to the container
leaves the model open.