Add persist option to ModelContainer
The default behavior of the ModelContainer iterator is to open one
image and close the previously opened image. There are some places in
the jwst code where it is assumed that all images remain open. To
avoid needing to rewrite this code I have added the persist keyword to
the ModelContainer __init__. If it is set to True (False is the
default) models will not be closed after they are opened.

I have also reviewed the jwst code and added the persist option in
those places where it has looked necessary. Also note that it is
possible to loop over models in a ModelContainer without using the
iterator and in this case models will not be closed either

    for i in range(len(container)):
        model = container[i]