Always create the EXTVER keyword when writing an hdu in datamodels
The code which writes EXTVER is controlled by the sequence_index field
in the schema. The default value (in the code, not the schema) is
None, which means EXTVER is not written if squence_index is not in the
schema. Setting the default to zero causes EXTVER to be set to one when
sequence_index is not in the schema.