Create WCS objects for the 1-D extracted spectra.
For each 1-D extracted spectrum, a WCS object is created that includes the
right ascension and declination at the nominal center of the slit, plus
the wavelength array that is also stored as a column in the output table.
Such a WCS is currently assigned individually to each output spectrum,
rather than creating a global WCS for model.meta.wcs.  GitHub issue #1034.