Modify datamodels to read and write wcs from/to SCI extension
The pipeline poobahs have decreed that the fits keywords describing
the wcs should be in the science extension instead the primary
header. To support this, the core schema has been updated with two
sets of wcs keywords: wcsinfo, which now points to the SCI extension,
and wcsinfo_primary, which points to the primary header. The code
which reads fits headers in fits_support.py has been modified to copy
wcs metadata from wcsinfo_primary to wcsinfo, so by the time the user
sees the metadata the wcs fields are under meta.wcsinfo. This copying
is only done if a fits file has a single SCI extension, none or more
than one leaves the wcs metadata under wcsinfo_primary. The
get_fits_wcs and set_fits_wcs methods now default to the SCI
extension. This shouldn't matter, as no code seems to use them.

This change will break existing tests. First, because any code that
writes a model will now write the wcs keywords to the SCI extension
and not to the primary header. Second, because datamodels now reads
the wcs in the SCI extension, the keywords there must validate. For
those reasons the tests and test data for datamodels was changed.