Roll back changes to fits_support and core schema
At the suggestion of the group, the code in fits_support no longer
copies the wcsinfo keuwords from the primary header to the science
extension. The core schema has been updated so it only loads wcsinfo
from the SCI extension. This means any wcs keywords in the primary
header are placed into extra_fits and are not directly accessible.

The method get_fits_wcs has been updated by adding the optional
keyword hdu_ver, which fetches the wcs from the specified version of
an extension.