Move wcsinfo to science extension (#1123)
This update will change test results, as wcs keywords will now be written to the SCI extension. I've updated the jwst tests where there were problems, but Pandokia results will need to be ok'd after being checked. If wcs keywords remain in the primary header, they will be duplicated, as datamodels makes sure that keywords not in the schema are copied. So when a new datamodel is copied from on old the wcs keywords will wind up in both the primary header and SCI extension.

The code which writes EXTVER is controlled by the sequence_index field
in the schema. The default value (in the code, not the schema) is
None, which means EXTVER is not written if squence_index is not in the
schema. Setting the default to zero causes EXTVER to be set to one when
sequence_index is not in the schema.

* Move wcsinfo to science extension

The core schema has been updated to read and write the wcs from the
science extension instead of the primary header. Tests and test data
have been updated to match. Default arguments of get_fits_wcs and
set_fits_wcs have been updated. A hdu_ver argument has been added to
get_fits_wcs.