Add fits_required to schema verification
The new field fits_required can be added to a schema. When it is
present, the data named in the schema must be present in the fits
file. The data can either be a keyword or extension. Normally the code
will throw a validation error if the data is missing, but if
pass_invalid_value is true, it will give a warning instead.