Add fits_required to schema verification (#1127)
The schemas are unchanged, so if you want to take advantage of this, add "fits_required: true" at the appropriate place in your schemas.

The code which writes EXTVER is controlled by the sequence_index field
in the schema. The default value (in the code, not the schema) is
None, which means EXTVER is not written if squence_index is not in the
schema. Setting the default to zero causes EXTVER to be set to one when
sequence_index is not in the schema.

* Move wcsinfo to science extension

The core schema has been updated to read and write the wcs from the
science extension instead of the primary header. Tests and test data
have been updated to match. Default arguments of get_fits_wcs and
set_fits_wcs have been updated. A hdu_ver argument has been added to
get_fits_wcs.

* Add fits_required to schema verification

The new field fits_required can be added to a schema. When it is
present, the data named in the schema must be present in the fits
file. The data can either be a keyword or extension. Normally the code
will throw a validation error if the data is missing, but if
pass_invalid_value is true, it will give a warning instead.